+++ src/N98/Magento/Application/Console/Events.php
@@ -1,0 +1,11 @@
+<?php
+
+namespace N98\Magento\Application\Console;
+
+final class Events
+{
+    /**
+     * @var string
+     */
+    const RUN_BEFORE = 'n98-magerun.application.console.run.before';
+}